+++ MekHQ/src/mekhq/campaign/parts/MissingProtomekJumpJet.java
@@ -42,7 +42,7 @@ public class MissingProtomekJumpJet extends MissingPart {
         this(0, null);
     }
     
-    public MissingProtomekJumpJet(int tonnage, Campaign c) {
+    public MissingProtomekJumpJet(double tonnage, Campaign c) {
         super(c);
         this.name = "Protomech Jump Jet"; //$NON-NLS-1$
         get(Installable.class).setLocations(Protomech.LOC_TORSO);
@@ -51,14 +51,14 @@ public class MissingProtomekJumpJet extends MissingPart {
     }
     
     @Override 
-	public int getBaseTime() {
-		return 60;
-	}
-	
-	@Override
-	public int getDifficulty() {
-		return 0;
-	}
+    public int getBaseTime() {
+        return 60;
+    }
+    
+    @Override
+    public int getDifficulty() {
+        return 0;
+    }
    
     @Override
     public double getTonnage() {
@@ -120,9 +120,9 @@ public class MissingProtomekJumpJet extends MissingPart {
     @Override
     public String checkFixable() {
         Unit unit = get(Installable.class).getUnit();
-    	if(null == unit) {
-    		return null;
-    	}
+        if(null == unit) {
+            return null;
+        }
         if(unit.isLocationBreached(Protomech.LOC_TORSO)) {
             return unit.getEntity().getLocationName(Protomech.LOC_TORSO) + " is breached.";
         }
@@ -175,19 +175,19 @@ public class MissingProtomekJumpJet extends MissingPart {
         return damagedJJ;
     }
 
-	@Override
-	public int getIntroDate() {
-		return 3055;
-	}
+    @Override
+    public int getIntroDate() {
+        return 3055;
+    }
 
-	@Override
-	public int getExtinctDate() {
-		return EquipmentType.DATE_NONE;
-	}
+    @Override
+    public int getExtinctDate() {
+        return EquipmentType.DATE_NONE;
+    }
 
-	@Override
-	public int getReIntroDate() {
-		return EquipmentType.DATE_NONE;
-	}
+    @Override
+    public int getReIntroDate() {
+        return EquipmentType.DATE_NONE;
+    }
 
 }
